+++ packages/app/components/comments/index.tsx
@@ -17,10 +17,7 @@ import { ListRenderItemInfo } from "@shopify/flash-list";
 
 import { useAlert } from "@showtime-xyz/universal.alert";
 import { useIsDarkMode } from "@showtime-xyz/universal.hooks";
-import {
-  InfiniteScrollList,
-  InfiniteScrollListV2,
-} from "@showtime-xyz/universal.infinite-scroll-list";
+import { InfiniteScrollList } from "@showtime-xyz/universal.infinite-scroll-list";
 import { useSafeAreaInsets } from "@showtime-xyz/universal.safe-area";
 import { colors } from "@showtime-xyz/universal.tailwind";
 import { View } from "@showtime-xyz/universal.view";
@@ -197,7 +194,7 @@ export function Comments({ nft, webListHeight }: CommentsProps) {
         <CommentsStatus isLoading={isLoading} error={error} />
       ) : (
         <View tw="web:pt-4 flex-grow">
-          <InfiniteScrollListV2
+          <InfiniteScrollList
             data={dataReversed}
             refreshing={isLoading}
             renderItem={renderItem}
