+++ packages/design-system/infinite-scroll-list/index.ts
@@ -1,2 +1,1 @@
 export * from "./infinite-scroll-list";
-export * from "./infinite-scroll-list-v2";
