--- packages/design-system/infinite-scroll-list/infinite-scroll-list-v2.tsx
@@ -1,1 +1,0 @@
-export { InfiniteScrollList as InfiniteScrollListV2 } from "./infinite-scroll-list";
