+++ packages/design-system/infinite-scroll-list/infinite-scroll-list.web.tsx
@@ -1,42 +1,240 @@
-import React, {
-  CSSProperties,
-  useMemo,
-  useRef,
+import {
   useEffect,
-  useState,
+  useRef,
   MutableRefObject,
+  useLayoutEffect,
+  isValidElement,
 } from "react";
 
 import type { FlashListProps, ViewToken } from "@shopify/flash-list";
-import type {
-  GridListProps,
-  VirtuosoHandle,
-  VirtuosoGridProps,
-  VirtuosoGridHandle,
-  GridComponents,
-  GridItemProps,
-  Components,
-} from "react-virtuoso";
-import { VirtuosoGrid, Virtuoso } from "react-virtuoso";
+import {
+  useWindowVirtualizer,
+  useVirtualizer,
+  Virtualizer,
+} from "@tanstack/react-virtual";
+import { useRouter } from "next/router";
 
-import type { InfiniteScrollListProps } from ".";
+import { useLatestValueRef } from "app/hooks/use-latest-value-ref";
+import { useStableCallback } from "app/hooks/use-stable-callback";
+import { delay } from "app/utilities";
 
 const DEFAULT_VIEWABILITY_THRESHOLD_PERCENTAGE = 80;
 
-export type InfiniteScrollListWebProps<T> = Omit<
-  InfiniteScrollListProps<T>,
-  "onEndReached"
-> & {
-  onEndReached?: VirtuosoGridProps<T>["endReached"];
-  onViewableItemsChanged?: FlashListProps<T>["onViewableItemsChanged"];
-  gridItemProps?: React.HTMLAttributes<HTMLDivElement>;
-};
-
 const renderComponent = (Component: any) => {
   if (!Component) return null;
-  if (React.isValidElement(Component)) return Component;
+  if (isValidElement(Component)) return Component;
   return <Component />;
 };
+function InfiniteScrollListV2Impl<Item>(
+  props: FlashListProps<Item> & {
+    useWindowScroll?: boolean;
+    overscan?: number;
+    preserveScrollPosition?: boolean;
+  }
+) {
+  const {
+    data,
+    renderItem,
+    extraData,
+    onViewableItemsChanged,
+    viewabilityConfig,
+    ItemSeparatorComponent,
+    estimatedItemSize,
+    ListHeaderComponent,
+    ListFooterComponent,
+    ListEmptyComponent,
+    onEndReached,
+    numColumns = 1,
+    style,
+    useWindowScroll = true,
+    preserveScrollPosition = false,
+  } = props;
+  let count = data?.length ?? 0;
+  if (numColumns) {
+    count = Math.ceil(count / numColumns);
+  }
+
+  const viewableItems = useRef<ViewToken[]>([]);
+  const parentRef = useRef<HTMLDivElement>(null);
+  const scrollMarginOffseRef = useRef<HTMLDivElement>(null);
+
+  const parentOffsetRef = useRef(0);
+  const router = useRouter();
+  const key = `myapp-scroll-restoration-${router.asPath}-window-scroll-${useWindowScroll}`;
+  const restoring = useRef(false);
+
+  useLayoutEffect(() => {
+    parentOffsetRef.current = scrollMarginOffseRef.current?.offsetTop ?? 0;
+  }, []);
+  let rowVirtualizer: Virtualizer<any, any>;
+  if (useWindowScroll) {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    rowVirtualizer = useWindowVirtualizer({
+      count,
+      estimateSize: () => estimatedItemSize ?? 0,
+      scrollMargin: parentOffsetRef.current,
+    });
+  } else {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    rowVirtualizer = useVirtualizer({
+      count,
+      estimateSize: () => estimatedItemSize ?? 0,
+      getScrollElement: () => parentRef.current,
+      scrollMargin: parentOffsetRef.current,
+    });
+  }
+
+  const scrollToLastSavedPosition = useStableCallback(async () => {
+    if (!preserveScrollPosition) return;
+    const pos = sessionStorage.getItem(key);
+    if (pos) {
+      const parsedPos = Number(pos);
+      const restoreScroll = async () => {
+        restoring.current = true;
+        rowVirtualizer.scrollToOffset(parsedPos);
+        // TODO: hack, not sure why scrollElement.scrollTo sometimes sets scrollY to < than parsedPos. I think it's because of height is less. Need to investigate
+        await delay(10);
+        if (rowVirtualizer.scrollOffset >= parsedPos) {
+          cancelAnimationFrame(req);
+          sessionStorage.removeItem(key);
+        } else {
+          requestAnimationFrame(restoreScroll);
+        }
+        restoring.current = false;
+      };
+
+      const req = requestAnimationFrame(restoreScroll);
+    }
+  });
+
+  const renderedItems = rowVirtualizer.getVirtualItems();
+
+  useEffect(() => {
+    const [lastItem] = [...renderedItems].reverse();
+    if (!lastItem) {
+      return;
+    }
+
+    if (data && lastItem.index >= data.length - 1) {
+      onEndReached?.();
+    }
+  }, [data, onEndReached, renderedItems]);
+
+  useEffect(() => {
+    scrollToLastSavedPosition();
+  }, [scrollToLastSavedPosition]);
+
+  useEffect(() => {
+    if (restoring.current || !preserveScrollPosition) return;
+    sessionStorage.setItem(key, rowVirtualizer.scrollOffset.toString());
+  }, [key, rowVirtualizer.scrollOffset, preserveScrollPosition]);
+  return (
+    <>
+      {useWindowScroll && renderComponent(ListHeaderComponent)}
+      <div
+        style={{
+          flex: 1,
+          display: "flex",
+          flexDirection: "column",
+        }}
+      >
+        {data?.length === 0 && renderComponent(ListEmptyComponent)}
+        <div
+          ref={parentRef}
+          style={
+            !useWindowScroll
+              ? {
+                  display: "flex",
+                  flexDirection: "column",
+                  overflowY: "auto",
+                  contain: "strict",
+                  //@ts-ignore
+                  ...style,
+                }
+              : {}
+          }
+        >
+          {!useWindowScroll && renderComponent(ListHeaderComponent)}
+          <div
+            ref={scrollMarginOffseRef}
+            style={{
+              height: rowVirtualizer.getTotalSize(),
+              width: "100%",
+              position: "relative",
+            }}
+          >
+            <div
+              style={{
+                position: "absolute",
+                top: 0,
+                left: 0,
+                width: "100%",
+                transform: `translateY(${
+                  renderedItems[0]?.start - rowVirtualizer.options.scrollMargin
+                }px)`,
+              }}
+            >
+              {renderedItems.map((virtualItem) => {
+                return (
+                  <div
+                    key={virtualItem.key}
+                    data-index={virtualItem.index}
+                    ref={rowVirtualizer.measureElement}
+                    style={{ width: "100%" }}
+                  >
+                    {typeof data?.[virtualItem.index] !== "undefined" ? (
+                      <div
+                        style={{
+                          display: "flex",
+                          width: "100%",
+                          justifyContent: "space-between",
+                        }}
+                      >
+                        {data
+                          .slice(
+                            virtualItem.index * numColumns,
+                            virtualItem.index * numColumns + numColumns
+                          )
+                          .map((item, i) => {
+                            const realIndex =
+                              virtualItem.index * numColumns + i;
+                            return (
+                              <ViewabilityTracker
+                                key={realIndex}
+                                index={realIndex}
+                                itemVisiblePercentThreshold={
+                                  viewabilityConfig?.itemVisiblePercentThreshold ??
+                                  DEFAULT_VIEWABILITY_THRESHOLD_PERCENTAGE
+                                }
+                                item={data[realIndex]}
+                                viewableItems={viewableItems}
+                                onViewableItemsChanged={onViewableItemsChanged}
+                              >
+                                {renderItem?.({
+                                  index: realIndex,
+                                  item,
+                                  extraData,
+                                  target: "Cell",
+                                }) ?? null}
+                                {realIndex < data.length - 1 &&
+                                  renderComponent(ItemSeparatorComponent)}
+                              </ViewabilityTracker>
+                            );
+                          })}
+                      </div>
+                    ) : null}
+                  </div>
+                );
+              })}
+            </div>
+          </div>
+          {!useWindowScroll && renderComponent(ListFooterComponent)}
+        </div>
+      </div>
+      {useWindowScroll && renderComponent(ListFooterComponent)}
+    </>
+  );
+}
 
 const ViewabilityTracker = ({
   index,
@@ -57,50 +255,54 @@ const ViewabilityTracker = ({
 
   useEffect(() => {
     let observer: IntersectionObserver;
-    // defer with a setTimeout. I think virtuoso might be mounting stuff async so intersection observer doesn't detect item on initial render
-    setTimeout(() => {
-      if (onViewableItemsChanged) {
-        observer = new IntersectionObserver(
-          ([entry]) => {
-            if (entry.isIntersecting) {
-              if (!viewableItems.current.find((v) => v.index === index))
-                viewableItems.current.push({
-                  item,
-                  index,
-                  isViewable: true,
-                  key: index.toString(),
-                  timestamp: new Date().valueOf(),
-                });
-            } else {
-              viewableItems.current = viewableItems.current.filter(
-                (v) => v.index !== index
-              );
-            }
-
-            viewableItems.current = viewableItems.current.sort((a, b) =>
-              a.index && b.index ? a.index - b.index : -1
+    if (onViewableItemsChanged) {
+      observer = new IntersectionObserver(
+        ([entry]) => {
+          if (entry.isIntersecting) {
+            if (
+              viewableItems.current.findIndex((v) => v.index === index) === -1
+            )
+              viewableItems.current.push({
+                item,
+                index,
+                isViewable: true,
+                key: index.toString(),
+                timestamp: new Date().valueOf(),
+              });
+          } else {
+            viewableItems.current = viewableItems.current.filter(
+              (v) => v.index !== index
             );
-
-            onViewableItemsChanged?.({
-              viewableItems: viewableItems.current,
-
-              // TODO: implement changed
-              changed: [],
-            });
-          },
-
-          {
-            // will trigger intersection callback when item is 70% visible
-            threshold: itemVisiblePercentThreshold / 100,
           }
-        );
 
-        if (ref.current) observer.observe(ref.current);
-      }
-    }, 10);
+          viewableItems.current = viewableItems.current.sort(
+            (a, b) =>
+              //@ts-ignore
+              a.index - b.index
+          );
+
+          onViewableItemsChanged?.({
+            viewableItems: viewableItems.current,
+
+            // TODO: implement changed
+            changed: [],
+          });
+        },
+
+        {
+          // will trigger intersection callback when item is 70% visible
+          threshold: itemVisiblePercentThreshold / 100,
+        }
+      );
+
+      if (ref.current) observer.observe(ref.current);
+    }
 
     return () => {
       observer?.disconnect();
+      viewableItems.current = viewableItems.current.filter(
+        (v) => v.index !== index
+      );
     };
   }, [
     onViewableItemsChanged,
@@ -110,180 +312,13 @@ const ViewabilityTracker = ({
     itemVisiblePercentThreshold,
   ]);
 
-  return <div ref={ref}>{children}</div>;
+  return (
+    <div style={{ width: "100%" }} ref={ref}>
+      {children}
+    </div>
+  );
 };
 
-export function VirtuosoListComponent<T>(
-  {
-    renderItem,
-    data,
-    onEndReached,
-    ListHeaderComponent,
-    ListFooterComponent,
-    ItemSeparatorComponent,
-    ListEmptyComponent,
-    numColumns = 1,
-    overscan,
-    useWindowScroll = true,
-    style,
-    onViewableItemsChanged,
-    gridItemProps = {},
-    viewabilityConfig,
-    estimatedItemSize,
-  }: InfiniteScrollListWebProps<T>,
-  ref: React.Ref<VirtuosoHandle> | React.Ref<VirtuosoGridHandle>
-) {
-  const viewableItems = useRef<ViewToken[]>([]);
-  const [listItemHeight, setListItemHeight] = useState(0);
+const InfiniteScrollList = InfiniteScrollListImpl;
 
-  const minHeight = useMemo(() => {
-    if (listItemHeight) {
-      return listItemHeight;
-    }
-    if (estimatedItemSize && data) {
-      return (
-        estimatedItemSize *
-        (numColumns ? data.length / numColumns : data.length)
-      );
-    }
-    return 0;
-  }, [data, estimatedItemSize, listItemHeight, numColumns]);
-
-  const renderItemContent = React.useCallback(
-    (index: number) => {
-      if (data && data[index]) {
-        const element = renderItem?.({
-          item: data[index],
-          index,
-          target: "Cell",
-        });
-        return (
-          <ViewabilityTracker
-            index={index}
-            itemVisiblePercentThreshold={
-              viewabilityConfig?.itemVisiblePercentThreshold ??
-              DEFAULT_VIEWABILITY_THRESHOLD_PERCENTAGE
-            }
-            item={data[index]}
-            viewableItems={viewableItems}
-            onViewableItemsChanged={onViewableItemsChanged}
-          >
-            {element}
-            {index < data.length - 1 && renderComponent(ItemSeparatorComponent)}
-          </ViewabilityTracker>
-        );
-      }
-
-      return null;
-    },
-    [
-      data,
-      ItemSeparatorComponent,
-      renderItem,
-      onViewableItemsChanged,
-      viewabilityConfig?.itemVisiblePercentThreshold,
-    ]
-  );
-
-  const renderListFooterComponent = React.useCallback(
-    () => renderComponent(ListFooterComponent),
-    [ListFooterComponent]
-  );
-  const components = useMemo<Components>(
-    () => ({
-      Header: ListHeaderComponent,
-      Footer: renderListFooterComponent,
-      EmptyPlaceholder: () => renderComponent(ListEmptyComponent),
-    }),
-    [ListEmptyComponent, ListHeaderComponent, renderListFooterComponent]
-  );
-  const gridComponents = useMemo<GridComponents<any>>(
-    () => ({
-      Item: (props: GridItemProps) => (
-        <ItemContainer
-          {...props}
-          numColumns={numColumns}
-          ItemSeparatorComponent={ItemSeparatorComponent}
-          {...(gridItemProps as {})}
-        />
-      ),
-      List: ListContainer,
-      Header: ListHeaderComponent,
-      Footer: renderListFooterComponent,
-    }),
-    [
-      ItemSeparatorComponent,
-      ListHeaderComponent,
-      gridItemProps,
-      numColumns,
-      renderListFooterComponent,
-    ]
-  );
-  return (
-    <div
-      style={
-        useWindowScroll ? { minHeight: `${minHeight}px` } : { height: "100%" }
-      }
-    >
-      {numColumns === 1 ? (
-        <Virtuoso
-          useWindowScroll={useWindowScroll}
-          data={data ?? []}
-          defaultItemHeight={estimatedItemSize}
-          endReached={onEndReached}
-          itemContent={renderItemContent}
-          components={components}
-          overscan={overscan}
-          style={style as CSSProperties}
-          totalListHeightChanged={setListItemHeight}
-          ref={ref as React.Ref<VirtuosoHandle>}
-        />
-      ) : (
-        <VirtuosoGrid
-          useWindowScroll={useWindowScroll}
-          totalCount={data?.length || 0}
-          components={gridComponents}
-          endReached={onEndReached}
-          itemContent={renderItemContent}
-          overscan={overscan}
-          style={style as CSSProperties}
-          ref={ref as React.Ref<VirtuosoGridHandle>}
-        />
-      )}
-    </div>
-  );
-}
-export const InfiniteScrollList = React.forwardRef(VirtuosoListComponent);
-
-const ListContainer = React.forwardRef(function ListContainer(
-  props: GridListProps,
-  ref: React.LegacyRef<HTMLDivElement>
-) {
-  return (
-    <div
-      {...props}
-      style={{ ...props.style, display: "flex", flexWrap: "wrap" }}
-      ref={ref}
-    />
-  );
-});
-
-type ItemContainerProps = GridItemProps &
-  Pick<FlashListProps<any>, "ItemSeparatorComponent"> & {
-    style?: CSSProperties;
-    numColumns?: number;
-    children?: JSX.Element;
-  } & React.HTMLAttributes<HTMLDivElement>;
-
-const ItemContainer = React.forwardRef(function ItemContainer(
-  { style, ...rest }: ItemContainerProps,
-  ref: React.LegacyRef<HTMLDivElement>
-) {
-  const width = rest.numColumns ? `${100 / rest.numColumns}%` : "100%";
-  return (
-    <div {...rest} style={{ ...style, width }} ref={ref}>
-      {rest.children}
-      {renderComponent(rest.ItemSeparatorComponent)}
-    </div>
-  );
-});
+export { InfiniteScrollList };
